warning handling
